move query from chain to token instance
