+++ apps/explorer/lib/explorer/chain.ex
@@ -4866,52 +4866,6 @@ defmodule Explorer.Chain do
      |> Enum.reduce(initial, reducer)}
   end
 
-  def unfetched_erc_721_token_instances_count do
-    nft_tokens =
-      from(
-        token in Token,
-        where: token.type == ^"ERC-721",
-        select: token.contract_address_hash
-      )
-
-    query =
-      from(
-        token_transfer in TokenTransfer,
-        inner_join: token in subquery(nft_tokens),
-        on: token.contract_address_hash == token_transfer.token_contract_address_hash,
-        left_join: instance in Instance,
-        on:
-          token_transfer.token_contract_address_hash == instance.token_contract_address_hash and
-            token_transfer.token_id == instance.token_id,
-        where: is_nil(instance.token_id) and not is_nil(token_transfer.token_id),
-        select: %{
-          contract_address_hash: token_transfer.token_contract_address_hash,
-          token_id: token_transfer.token_id
-        }
-      )
-
-    distinct_query =
-      from(
-        q in subquery(query),
-        distinct: [q.contract_address_hash, q.token_id]
-      )
-
-    count_query =
-      from(
-        c in subquery(distinct_query),
-        select: %{
-          count: fragment("COUNT(*)")
-        }
-      )
-
-    result = Repo.one(count_query)
-
-    case result do
-      nil -> 0
-      row -> row.count
-    end
-  end
-
   @doc """
   Streams a list of token contract addresses that have been cataloged.
   """
